# Use jagged instead of multidimensional array