--- a/src/ImageProcessorCore/Samplers/Processors/Convolution/EdgeDetection/RobertsCrossProcessor.cs
+++ b/src/ImageProcessorCore/Samplers/Processors/Convolution/EdgeDetection/RobertsCrossProcessor.cs
@@ -15,18 +15,22 @@ namespace ImageProcessorCore.Processors
         where TColor : IPackedVector<TPacked>
         where TPacked : struct
     {
-        /// <inheritdoc/>
-        public override float[,] KernelX => new float[,]
+        private static readonly float[][] kernelX = new float[2][]
         {
-            { 1, 0 },
-            { 0, -1 }
+            new float[] { 1, 0 },
+            new float[] { 0, -1 }
+        };
+
+        private static readonly float[][] kernelY = new float[2][]
+        {
+            new float[] { 0, 1 },
+            new float[] { -1, 0 }
         };
 
         /// <inheritdoc/>
-        public override float[,] KernelY => new float[,]
-        {
-            { 0, 1 },
-            { -1, 0 }
-        };
+        public override float[][] KernelX => kernelX;
+
+        /// <inheritdoc/>
+        public override float[][] KernelY => kernelY;
     }
 }