--- a/src/ImageProcessorCore/Samplers/Processors/Convolution/EdgeDetection/SobelProcessor.cs
+++ b/src/ImageProcessorCore/Samplers/Processors/Convolution/EdgeDetection/SobelProcessor.cs
@@ -15,20 +15,24 @@ namespace ImageProcessorCore.Processors
         where TColor : IPackedVector<TPacked>
         where TPacked : struct
     {
-        /// <inheritdoc/>
-        public override float[,] KernelX => new float[,]
+        private static readonly float[][] kernelX = new float[3][]
         {
-            { -1, 0, 1 },
-            { -2, 0, 2 },
-            { -1, 0, 1 }
+            new float[] { -1, 0, 1 },
+            new float[] { -2, 0, 2 },
+            new float[] { -1, 0, 1 }
+        };
+
+        private static readonly float[][] kernelY = new float[3][]
+        {
+            new float[] { -1, -2, -1 },
+            new float[] { 0, 0, 0 },
+            new float[] { 1, 2, 1 }
         };
 
         /// <inheritdoc/>
-        public override float[,] KernelY => new float[,]
-        {
-            { -1, -2, -1 },
-            { 0, 0, 0 },
-            { 1, 2, 1 }
-        };
+        public override float[][] KernelX => kernelX;
+
+        /// <inheritdoc/>
+        public override float[][] KernelY => kernelY;
     }
 }