Use jagged instead of multidimensional array as suggested in #462.
Only create the kernels once in the filters.