@dlemstra 2016-10-23
I added the benchmarks and I am getting the following results on my machine (2 runs):

_before:_
```
|      Median |    StdDev | Gen 0 | Gen 1 |  Gen 2 | Bytes Allocated/Op |
|------------ |---------- |------ |------ |------- |------------------- |
| 426.0806 ms | 4.7395 ms |     - |     - | 450.00 |       8.008.666,72 |
| 427.2504 ms | 4.7152 ms |     - |     - | 496.00 |       8.338.740,79 |
```

_after:_
```
|      Median |    StdDev | Gen 0 | Gen 1 |  Gen 2 | Bytes Allocated/Op |
|------------ |---------- |------ |------ |------- |------------------- |
| 411.7357 ms | 3.7951 ms |     - |     - | 446.00 |       8.012.231,19 |
| 412.0662 ms | 4.3176 ms |     - |     - | 449.00 |       7.873.010,86 |
```
So this is only 3.6% faster but I think we should go ahead and replace the other multidimensional arrays with jagged arrays. I can do that after this one has been merged.

Can you clarify what you mean by double checking the properties and where am I not using C#6?