@JimBobSquarePants 2016-10-23
>Can you clarify what you mean by double checking the properties and where am I not using C#6?

Yeah, instead of using `private static` etc you should be able to do something like 

`public override float[][] KernelX { get; } = { */..array here..*/}`