@dlemstra 2016-10-23
I changed it to this so we would not create a new array each time. The C#6 syntax would create a new array each time but the array is never modified.