@JimBobSquarePants 2016-10-23
`=>` would create a new array. I'm fairly certain simply using `=` doesn't though.