@dlemstra 2016-10-23
It won't for this instance but it will for every new instance that is created. I think we can just share the same array between all instances and threads.